storage/requirements.txt
fastapi
uvicorn
gunicorn
python-decouple
pydantic
aiofiles
